Minor changes after review
